# perf: simplify encoder's LRU