LRU implementation is extracted from inside Encoder to a new **lru** struct and it is now simplified for easier maintainability.  It easier to read, test and benchmark. The sec/op is pretty much similar but the B/op and allocs/op is slightly reduced.

benchstat:
```js
goos: darwin
goarch: amd64
pkg: github.com/muktihari/fit/encoder
cpu: Intel(R) Core(TM) i5-5257U CPU @ 2.70GHz
                                     │   old.txt    │              new.txt               │
                                     │    sec/op    │   sec/op     vs base               │
Encode/normal_header_zero-4            80.07m ±  7%   79.36m ± 4%       ~ (p=0.063 n=10)
Encode/normal_header_15-4              81.42m ± 73%   80.19m ± 8%       ~ (p=0.247 n=10)
Encode/compressed_timestamp_header-4   72.55m ± 30%   72.38m ± 0%       ~ (p=0.123 n=10)
geomean                                77.91m         77.23m       -0.88%

                                     │   old.txt    │               new.txt               │
                                     │     B/op     │     B/op      vs base               │
Encode/normal_header_zero-4            3.304Mi ± 0%   3.133Mi ± 0%  -5.18% (p=0.000 n=10)
Encode/normal_header_15-4              3.054Mi ± 0%   3.053Mi ± 0%       ~ (p=0.053 n=10)
Encode/compressed_timestamp_header-4   2.517Mi ± 0%   2.342Mi ± 0%  -6.96% (p=0.000 n=10)
geomean                                2.939Mi        2.819Mi       -4.10%

                                     │   old.txt   │              new.txt               │
                                     │  allocs/op  │  allocs/op   vs base               │
Encode/normal_header_zero-4            108.5k ± 0%   103.5k ± 0%  -4.59% (p=0.000 n=10)
Encode/normal_header_15-4              101.0k ± 0%   101.0k ± 0%  -0.02% (p=0.000 n=10)
Encode/compressed_timestamp_header-4   108.5k ± 0%   103.5k ± 0%  -4.59% (p=0.000 n=10)
geomean                                105.9k        102.7k       -3.09%

```